Fix  eslint  no-multi-comp
Move GalleryModal to it's own file
